fix warnings

+++ FABRIK/src/fabrik2D.cpp
@@ -135,8 +135,8 @@ void Fabrik2D::RandomizeTargets(int min, int max) {
 	int range = max - min + 1;
 
 	for (int i = 0; i < targets->size(); i++) {
-		float x = (rand() % range + min);
-		float y = (rand() % range + min);
+		float x = static_cast<float>(rand() % range + min);
+		float y = static_cast<float>(rand() % range + min);
 
 		Vector2 vector = { x, y };
 
@@ -189,7 +189,7 @@ void Fabrik2D::Forward() {
 
 		//Krok niepotrzebny, suma wekorow nowych pozycji subbase'a z kazdego lanucucha wystarczy do wyznaczenia kierunku, a dlugosc miedzy stawami i tak jest zachowana
 		if (nodeJoint == subbase) {
-			subbase->value.PositionTmp = subbase->value.PositionTmp / subbase->child.size();
+			subbase->value.PositionTmp = subbase->value.PositionTmp / (float)subbase->child.size();
 		}
 
 		if (nodeJoint->parent != tree->root && nodeJoint != tree->root) {
